Refactor the way an event's name is calculated, stored, and retrieved.
Use a single method on the model to access the first and second parts
of the event title (billing). When the event title is disabled, use the
headliner as the event title and supporting acts as the subtext. When
it's enabled, use the title as headline and all acts as subtitle.